backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.2
sqlalchemy==2.0.40
psycopg[binary]>=3.2.10
openai==1.75.0
pytesseract==0.3.13
Pillow==11.2.1
pgvector==0.3.6
